# PERF: operations with zoneinfo tzinfos