```
import zoneinfo
import pandas as pd

tz = zoneinfo.ZoneInfo("US/Pacific")
dti = pd.date_range("2016-01-01", periods=10**5, freq="s", tz=tz)

In [3]: %timeit dti.normalize()
142 ms ± 3.44 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- main
90.4 ms ± 4.34 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- PR
```

The corresponding timing with pytz is 49.9 ms.  @pganssle any thoughts on getting to near-parity (so we can drop pytz xref #46463)?  Is there any prospect of exposing zoneinfo_fromutc in the C-API?